Version 2.0
The major change in this version has to do with the outcome and
traversal encodings. Outcomes are seven bits shorter, and traversals are
twelve bits longer than before. This was achieved by only encoding the
deaths of optional allies in the outcome, since any required allies that
do not survive are, by definition, dead. In a similar way, encoded
loyalties mask off Morinth's bit, since she is always loyal. Telling the
user that Morinth is loyal or that they need to do her loyalty mission
is redundant. Loyalty is encoded in both the outcome and the traversal,
but the outcome loyalty is masked with surviving allies. This is because
outcomes that only differ by the loyalties of dead squadmates are not
actually different.

Morinth was incorrectly included in the choose_recruitment and
choose_recruits iterations. To fix this, we now distinguish between
optional and recruitable allies.