Version 3.0
This version reduces the average size of traversal encodings by taking
advantage of the fact that the specific ally selected as the first
fireteam leader is not necessary knowledge, as long as the user
understands what is meant by "ideal" and "non-ideal". That part of the
encoding dropped from 4 bits down to 0–1 bits, and the decision tree
logic now only checks if there *are* any ideal first fireteam leaders
(i.e., the maximum fan-out for choose_first_leader() is now 2 instead
of N).

Similarly, the "unpicks" for the long walk are moot if there are only
two remaining allies from which to choose, so they are not encoded in
that case. However, the traversal decoder would not know the size of the
pool, so an additional mandatory bit was added to the encoding. In the
best case, this still saves four bits, but the worst case is one bit
wider than before.

The savings from both changes still amounts to an overall reduction in
the encoding.