Remove Bits and hide encdec imports
Based on fairly damning evidence, my "Bits" subclass of int was
slowing down encoding/decoding by nearly a factor of 2. Bits also caused
some headache when it accidentally made its way into the data file,
particularly because its pickled size is significantly larger than int.
In the end, it was simply more trouble than it was worth.

This commit also hides some imported symbols in the encdec module by
prepending an underscore to the names. The next commit will finish this
refactor for defs and dt.