Version 4.0
Unfortunately, the assumption that any available ally could be selected
as a tech specialist was invalid, so possibly millions of traversals
were recorded that are impossible.

An opportunistic improvement is also included, which names the "ideal"
leaders that should (not) be selected for the second fireteam.
Descriptions of the teams the leaders lead are now standardized to
"second fireteam" and "diversion team" to avoid confusion.