Version 5.0
After some thought (and watching several YouTube videos), I realized
that there is no significant difference between a dead ally and an
unrecruited ally when carried over to ME3 (except, possibly, some
dialog alluding to their death?). This means the "dead" bits in the
outcome were essentially shifted to the traversal encoding. Instead of
rerunning the entire thing, I decided to post-process the data into the
new format and modify the code to reflect the change.

Overall, it took about three weeks to finally produce the intended data
set, which, on disk, is ~13.5 MB. Over 39.6 billion traversals were
counted, which is significantly below my 20-trillion-ish estimate. :)